+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,2 +1,3 @@
 export { keyringController } from './keyring-controller';
+export { remoteFeatureFlagController } from './remote-feature-flag-controller';
 export { storageService } from './storage-service';